If `maven.repo.uri` has been redeclared by user, automatically set allow insecure http downloads.
